+++ probatus/feature_elimination/__init__.py
@@ -1,3 +1,4 @@
-from .feature_elimination import ShapRFECV, EarlyStoppingShapRFECV
+from .feature_elimination import ShapRFECV
+from .early_stopping_feature_elimination import EarlyStoppingShapRFECV
 
 __all__ = ["ShapRFECV", "EarlyStoppingShapRFECV"]
